feat(logstream): implement the LogStream append-only log (#57)
* feat(logstream): implement the LogStream append-only log

Implement docs/design/wal.md. Append claims a sequence number with one
conditional create, so no coordination service is needed. A gate probe at
seq+16 tells head contention from drift, and a gallop closes a drift in
about 2*log2(delta) probes. Tail reads one list page, then gallops.

The binary search is hand-written. sort.Search carries no context and
returns no error, so a failed probe cannot stop it, and it counts in
signed int, which cannot hold the upper half of the uint64 sequence
domain. wal.md records that reasoning.

gallop_internal_test.go is an internal test package. It reaches the
unexported gallop to prove the two constraints above: the context abort
and a head above math.MaxInt64. Every other test uses logstream_test.

* refactor(logstream,recordio): apply review, and use sort.Search

Rename gallop to jump. Lift the search into two free functions that take
the probe as a parameter, so a test drives them with no store. The binary
phase now uses sort.Search over the offset, where the width is bounded by
the real distance and always fits an int. The closure short-circuits on a
probe error or a canceled context, so a failure costs no round trip.

Name the per-stream channel ch, because a sync.Mutex cannot observe a
context and the field is not a mutex. Drop the CONSIDER comment that
repeated wal.md. Drop the timer.Stop call.

Carry the record-size limit as int, in recordio too. A record is a byte
slice, and its length is an int, so a wider limit names a record that
cannot exist. The frame format does not change, and the header still
holds the length as a little-endian uint64. Both readers reject a
non-positive limit, so a negative value cannot widen to a large uint64.

* chore: drop the petr review transcript

* refactor(logstream,recordio): apply code review findings

Inline the binary search loop into jump to eliminate integer overflow.
Drop the sort dependency and streamPrefix helper. Delete existing GCS
objects before running TestGCS and during cleanup. Fold concurrent
append tests into one table-driven test. Test lock contention
cancellation in jump_internal_test. Drop the unreachable non-positive
checks in recordio readers.

+++ recordio/crc.go
@@ -6,8 +6,8 @@ import (
 )
 
 const (
-	DefaultBufferSize          = 64 * 1024
-	DefaultMaxRecordSize int64 = 64 * 1024 * 1024
+	DefaultBufferSize        = 64 * 1024
+	DefaultMaxRecordSize int = 64 * 1024 * 1024
 
 	maskDelta     uint32 = 0xa282ead8
 	headerSize           = 12
